remove unuse components and rename PageFormSelect
