+++ framework/index.ts
@@ -31,11 +31,9 @@ export * from './PageForm/Inputs/FormGroupTextInput';
 export * from './PageForm/Inputs/PageFormCheckbox';
 export * from './PageForm/Inputs/PageFormDataEditor';
 export * from './PageForm/Inputs/PageFormSelect';
-export * from './PageForm/Inputs/PageFormSelectOption';
 export * from './PageForm/Inputs/PageFormSwitch';
 export * from './PageForm/Inputs/PageFormTextArea';
 export * from './PageForm/Inputs/PageFormTextInput';
-export * from './PageForm/Inputs/PageFormTextSelect';
 export * from './PageForm/PageForm';
 export * from './PageFramework';
 export * from './PageHeader';
